LayoutTests/fast/selectors/querySelector-multiple-simple-child-backtracking-expected.txt

Test backtracking with multiple chains of child relations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll("div>ul>li>div ul>li ul>li p").length is 1
PASS document.querySelectorAll("div>ul>li>div ul>li ul>li p")[0].id is "target"
PASS successfullyParsed is true

TEST COMPLETE
